+++ NEWS.md
@@ -1,4 +1,6 @@
-# secretbase 0.3.0.9000 (development)
+# secretbase 0.3.0.9001 (development)
+
+* Adds SipHash-1-3 pseudorandom function (PRF).
 
 # secretbase 0.3.0.1